terminal: Small changes

+++ base/new/include/hal/ports64.h
@@ -17,21 +17,19 @@ void out16 ( unsigned short port, unsigned short data );
 void out32 ( unsigned short port, unsigned int   data );
 
 
-
-
 void __x86_io_delay (void);
 
 void wait_ns(int count);
 
-//
-// para acesso via user mode.
-//
 
-unsigned int portsx86_IN ( int bits, unsigned int port );
+// Service 126.
+unsigned int portsx86_IN ( int bits, unsigned short port );
+
+// Service 127.
 void 
 portsx86_OUT ( 
     int bits, 
-    unsigned int port, 
+    unsigned short port, 
     unsigned int value );
