+++ com/lib/lib/libio01/include/libio.h
@@ -6,12 +6,12 @@
 
 
 // Retorna o valor.
-unsigned char  libio_inport8  (unsigned int port);
-unsigned short libio_inport16 (unsigned int port);
-unsigned long  libio_inport32 (unsigned int port);
+unsigned char  libio_inport8  (unsigned short port);
+unsigned short libio_inport16 (unsigned short port);
+unsigned int   libio_inport32 (unsigned short port);
 
-void libio_outport8  ( unsigned int port, unsigned char  value);
-void libio_outport16 ( unsigned int port, unsigned short value);
-void libio_outport32 ( unsigned int port, unsigned long  value);
+void libio_outport8  ( unsigned short port, unsigned char  value );
+void libio_outport16 ( unsigned short port, unsigned short value );
+void libio_outport32 ( unsigned short port, unsigned int   value );
